+++ transports/bluetooth/src/sys/hci_scan.rs
@@ -77,45 +77,6 @@ fn start_thread(sender: oneshot::Sender<Result<Vec<Addr>, io::Error>>) {
 }
 
 fn query() -> Result<Vec<Addr>, io::Error> {
-    unsafe {
-        let socket = libc::socket(libc::AF_BLUETOOTH, libc::SOCK_RAW | libc::SOCK_CLOEXEC, ffi::BTPROTO_HCI);
-        if socket == 0 {
-            return Err(io::Error::last_os_error());
-        }
-
-        let dev_id = ffi::hci_get_route(ptr::null_mut());
-        if dev_id == -1 {
-            libc::close(socket);
-            return Err(io::Error::last_os_error());
-        }
-
-        let num_results: u8 = 255;
-
-        let mut buf: Vec<u8> = Vec::with_capacity(mem::size_of::<ffi::hci_inquiry_req>() + mem::size_of::<ffi::inquiry_info>() * num_results as usize);
-        buf.set_len(buf.capacity());
-        let (req, results) = buf.split_at_mut(mem::size_of::<ffi::hci_inquiry_req>());
-        let mut req: *mut ffi::hci_inquiry_req = req.as_mut_ptr() as *mut _;
-        let results: *mut ffi::inquiry_info = results.as_mut_ptr() as *mut _;
-
-        (*req).dev_id = dev_id as u16;
-        (*req).flags = ffi::IREQ_CACHE_FLUSH as u16;
-        (*req).lap = [0x33, 0x8b, 0x9e];
-        (*req).length = 8;     // Timeout; the actual timeout is 1.28 times this value, don't ask me why
-        (*req).num_rsp = num_results;
-
-        let ret = libc::ioctl(socket, ffi::HCIINQUIRY, req as usize);
-        if ret < 0 {
-            libc::close(socket);
-            return Err(io::Error::last_os_error());
-        }
-
-        let mut out = Vec::with_capacity((*req).num_rsp as usize);
-        for elem in (0..(*req).num_rsp).map(|n| results.offset(n as isize)) {
-            let addr = Addr::from_little_endian((*elem).baddr.b);
-            out.push(addr);
-        }
-
-        libc::close(socket);
-        Ok(out)
-    }
+    let socket = HciSocket::new()?;
+    socket.inquiry()
 }
